8289710: Move Suspend/Resume classes out of os.hpp
