+++ src/hotspot/share/jfr/periodic/sampling/jfrThreadSampler.cpp
@@ -40,6 +40,7 @@
 #include "runtime/javaThread.inline.hpp"
 #include "runtime/os.hpp"
 #include "runtime/semaphore.hpp"
+#include "runtime/suspend.hpp"
 #include "runtime/threadCrashProtection.hpp"
 #include "runtime/threadSMR.hpp"
 
@@ -116,12 +117,12 @@ class JfrThreadSampleClosure {
   uint _added_native;
 };
 
-class OSThreadSampler : public os::SuspendedThreadTask {
+class OSThreadSampler : public SuspendedThreadTask {
  public:
   OSThreadSampler(JavaThread* thread,
                   JfrThreadSampleClosure& closure,
                   JfrStackFrame *frames,
-                  u4 max_frames) : os::SuspendedThreadTask((Thread*)thread),
+                  u4 max_frames) : SuspendedThreadTask((Thread*)thread),
     _success(false),
     _thread_oop(thread->threadObj()),
     _stacktrace(frames, max_frames),
@@ -129,8 +130,8 @@ class OSThreadSampler : public os::SuspendedThreadTask {
     _suspend_time() {}
 
   void take_sample();
-  void do_task(const os::SuspendedThreadTaskContext& context);
-  void protected_task(const os::SuspendedThreadTaskContext& context);
+  void do_task(const SuspendedThreadTaskContext& context);
+  void protected_task(const SuspendedThreadTaskContext& context);
   bool success() const { return _success; }
   const JfrStackTrace& stacktrace() const { return _stacktrace; }
 
@@ -144,7 +145,7 @@ class OSThreadSampler : public os::SuspendedThreadTask {
 
 class OSThreadSamplerCallback : public CrashProtectionCallback {
  public:
-  OSThreadSamplerCallback(OSThreadSampler& sampler, const os::SuspendedThreadTaskContext &context) :
+  OSThreadSamplerCallback(OSThreadSampler& sampler, const SuspendedThreadTaskContext &context) :
     _sampler(sampler), _context(context) {
   }
   virtual void call() {
@@ -152,10 +153,10 @@ class OSThreadSamplerCallback : public CrashProtectionCallback {
   }
  private:
   OSThreadSampler& _sampler;
-  const os::SuspendedThreadTaskContext& _context;
+  const SuspendedThreadTaskContext& _context;
 };
 
-void OSThreadSampler::do_task(const os::SuspendedThreadTaskContext& context) {
+void OSThreadSampler::do_task(const SuspendedThreadTaskContext& context) {
 #ifndef ASSERT
   guarantee(JfrOptionSet::sample_protection(), "Sample Protection should be on in product builds");
 #endif
@@ -177,7 +178,7 @@ void OSThreadSampler::do_task(const os::SuspendedThreadTaskContext& context) {
 * From this method and down the call tree we attempt to protect against crashes
 * using a signal handler / __try block. Don't take locks, rely on destructors or
 * leave memory (in case of signal / exception) in an inconsistent state. */
-void OSThreadSampler::protected_task(const os::SuspendedThreadTaskContext& context) {
+void OSThreadSampler::protected_task(const SuspendedThreadTaskContext& context) {
   JavaThread* const jt = JavaThread::cast(context.thread());
   // Skip sample if we signaled a thread that moved to other state
   if (!thread_state_in_java(jt)) {
