+++ src/hotspot/share/runtime/suspend.cpp
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 1997, 2022, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ *
+ */
+
+#include "precompiled.hpp"
+#include "runtime/atomic.hpp"
+#include "runtime/suspend.hpp"
+
+void SuspendedThreadTask::run() {
+  internal_do_task();
+  _done = true;
+}
+
+#ifndef _WINDOWS
+/* try to switch state from state "from" to state "to"
+ * returns the state set after the method is complete
+ */
+SuspendResume::State SuspendResume::switch_state(SuspendResume::State from,
+                                                 SuspendResume::State to)
+{
+  SuspendResume::State result = Atomic::cmpxchg(&_state, from, to);
+  if (result == from) {
+    // success
+    return to;
+  }
+  return result;
+}
+#endif
